requirements.txt
google-adk[database]==1.0.0
yfinance==0.2.56
python-dotenv==1.1.0
psutil==6.1.1
litellm==1.55.10
google-generativeai==0.21.1